refactor(recommend): refactor notification
