+++ ozpcenter/api/storefront/model_access.py
@@ -245,11 +245,11 @@ def get_recommendation_listing_ids(profile_instance):
             else:
                 recommendation_combined_dict['profile'][current_listing_id] = current_listing_score * recommender_name_weight
 
-    sorted_recommendations_combined_dict = recommend_utils.get_top_n_score(recommendation_combined_dict, 40)
+    sorted_recommendations_combined_dict = recommend_utils.get_top_n_score(recommendation_combined_dict['profile'], 40)
     # sorted_recommendations_combined_dict = {'profile': [[11, 8.5], [112, 8.0], [85, 7.0], [86, 7.0], [87, 7.0],
     #    [88, 7.0], [89, 7.0], [90, 7.0], [81, 6.0], [62, 6.0],
     #    [21, 5.5], [1, 5.0], [113, 5.0], [111, 5.0], [114, 5.0], [64, 4.0], [66, 4.0], [68, 4.0], [70, 4.0], [72, 4.0]]}
-    listing_ids_list = [entry[0] for entry in sorted_recommendations_combined_dict['profile']]
+    listing_ids_list = [entry[0] for entry in sorted_recommendations_combined_dict]
     return listing_ids_list, recommended_entry_data
